Move pseudo_connect to a new file.
